feat: major folder restructure
