+++ src/blog/__init__.py
@@ -1,0 +1,4 @@
+"""
+Flask Blog application rewritten using Hexagonal Architecture and with enterprise patterns.
+"""
+
