+++ src/blog/adapters/entrypoints/app/templates/auth/login.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends '../base.html' %}
 
 {% block header %}
   <h1>{% block title %}Log In{% endblock %}</h1>
