+++ src/blog/adapters/entrypoints/app/templates/auth/register.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends '../base.html' %}
 
 {% block header %}
   <h1>{% block title %}Register{% endblock %}</h1>
