+++ src/blog/adapters/entrypoints/app/templates/base.html
