+++ src/blog/adapters/entrypoints/app/templates/post/create.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends '../base.html' %}
 
 {% block header %}
   <h1>{% block title %}New Post{% endblock %}</h1>
