+++ src/blog/adapters/entrypoints/app/templates/post/index.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends '../base.html' %}
 
 {% block header %}
   <h1>{% block title %}Posts{% endblock %}</h1>
