+++ src/blog/adapters/repositories/user_repository.py
@@ -1,9 +1,9 @@
 from sqlite3 import Connection
-from src.domain.ports.repository import RepositoryInterface
+from src.blog.domain.ports.repositories.repository import RepositoryInterface
 from typing import Callable, Any
 
 
-class PostRepository(RepositoryInterface):
+class UserRepository(RepositoryInterface):
 
     def __init__(self, db_conn: Callable[[], Connection]) -> None:
         self.db_conn = db_conn()
